# docs/swagger.yaml
basePath: /
definitions:
  achievementmodel.Achievement:
    properties:
      description:
        type: string
      exp:
        description: количество экспы за открытие
        type: integer
      id:
        type: integer
      level:
        description: уровень (от 1 до 10)
        type: integer
      openDate:
        description: дата открытия ачивки по timestamp(0) в секундах, если достижение
          не открыто, равняется 0
        type: integer
      percentage:
        description: проценты до достижения (от 0 до 100), на открытых или достигнутых
          ачивках равняется 100
        type: integer
      reachDate:
        description: дата достижение пользователем ачивки по timestamp(0) в секундах,
          если достижение не достигнуто, равняется 0
        type: integer
      shown:
        description: была ли ачивка показана пользователю
        type: boolean
      type:
        $ref: '#/definitions/achievementmodel.AchievementType'
    type: object
  achievementmodel.AchievementType:
    enum:
    - Длительность
    - Сигареты
    - Здоровье
    - Самочувствие
    - Экономия
    type: string
    x-enum-varnames:
    - DURATION
    - CIGARETTE
    - HEALTH
    - WELL_BEING
    - SAVING
  errormodel.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  model.FloatUserStatistics:
    properties:
      day:
        type: number
      month:
        type: number
      week:
        type: number
      year:
        type: number
    type: object
  model.IntUserStatistics:
    properties:
      day:
        type: integer
      month:
        type: integer
      week:
        type: integer
      year:
        type: integer
    type: object
  model.ReachedAchievements:
    properties:
      achievementType:
        $ref: '#/definitions/achievementmodel.AchievementType'
      cigarette:
        type: integer
      duration:
        type: integer
      health:
        type: integer
      saving:
        type: integer
      well-being:
        type: integer
    type: object
...
  usermodel.PrivacySetting:
    enum:
    - STATISTICS_MONEY
    - STATISTICS_CIGARETTE
    - STATISTICS_LIFE
    - STATISTICS_TIME
    - ACHIEVEMENTS_DURATION
    - ACHIEVEMENTS_HEALTH
    - ACHIEVEMENTS_WELL_BEING
    - ACHIEVEMENTS_SAVING
    - ACHIEVEMENTS_CIGARETTE
    type: string
    x-enum-varnames:
    - STATISTICS_MONEY
    - STATISTICS_CIGARETTE
    - STATISTICS_LIFE
    - STATISTICS_TIME
    - ACHIEVEMENTS_DURATION
    - ACHIEVEMENTS_HEALTH
    - ACHIEVEMENTS_WELL_BEING
    - ACHIEVEMENTS_SAVING
    - ACHIEVEMENTS_CIGARETTE
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kurilka API Swagger
  version: "1.0"
paths:
  /achievements:
    get:
      description: return all achievements that exists, if user not reach achievement
        reachDate is zero, is user not open achievement openDate is zero
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/achievementmodel.Achievement'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: UserAchievements
      tags:
      - achievements
  /achievements/mark-shown:
    post:
      description: set on all reach achievements show = true
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: MarkShown
      tags:
      - achievements
  /achievements/open-single:
    post:
      description: open single achievement by user and achievement ids
      parameters:
      - description: achievement id
        in: query
        name: achievementId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: OpenSingle
      tags:
      - achievements
  /achievements/user-reached:
    get:
      description: get user reached achievements
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ReachedAchievements'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: UserReachedAchievements
      tags:
      - achievements
  /privacysettings:
    get:
      description: get user privacy settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/usermodel.PrivacySetting'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: GetPrivacySettingsHandler
      tags:
      - privacysettings
  /privacysettings/add:
    post:
      description: add one user privacy setting, if setting exists return error
      parameters:
      - description: privacy setting
        enum:
        - STATISTICS_MONEY
        - STATISTICS_CIGARETTE
        - STATISTICS_LIFE
        - STATISTICS_TIME
        - ACHIEVEMENTS_DURATION
        - ACHIEVEMENTS_HEALTH
        - ACHIEVEMENTS_WELL_BEING
        - ACHIEVEMENTS_SAVING
        - ACHIEVEMENTS_CIGARETTE
        in: query
        name: privacySetting
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: AddPrivacySetting
      tags:
      - privacysettings
  /privacysettings/remove:
    delete:
      description: remove one user privacy setting, if setting not exists return error
      parameters:
      - description: privacy setting
        enum:
        - STATISTICS_MONEY
        - STATISTICS_CIGARETTE
        - STATISTICS_LIFE
        - STATISTICS_TIME
        - ACHIEVEMENTS_DURATION
        - ACHIEVEMENTS_HEALTH
        - ACHIEVEMENTS_WELL_BEING
        - ACHIEVEMENTS_SAVING
        - ACHIEVEMENTS_CIGARETTE
        in: query
        name: privacySetting
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: RemovePrivacySetting
      tags:
      - privacysettings
  /privacysettings/switch:
    put:
      description: add privacy setting if not exists and delete if exists
      parameters:
      - description: privacy setting
        enum:
        - STATISTICS_MONEY
        - STATISTICS_CIGARETTE
        - STATISTICS_LIFE
        - STATISTICS_TIME
        - ACHIEVEMENTS_DURATION
        - ACHIEVEMENTS_HEALTH
        - ACHIEVEMENTS_WELL_BEING
        - ACHIEVEMENTS_SAVING
        - ACHIEVEMENTS_CIGARETTE
        in: query
        name: privacySetting
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: SwitchPrivacySettings
      tags:
      - privacysettings
  /statistics/cigarette:
    get:
      description: get cigarette statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntUserStatistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: CigaretteStatistics
      tags:
      - statistics
  /statistics/money:
    get:
      description: get money statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.FloatUserStatistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: MoneyStatistics
      tags:
      - statistics
  /statistics/time:
    get:
      description: get time statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntUserStatistics'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errormodel.ErrorResponse'
      summary: TimeStatistics
      tags:
      - statistics
...
swagger: "2.0"